Pull out logic from langchain integration
