+++ integrations/langchain/src/databricks_langchain/vector_search_retriever_tool.py
@@ -4,37 +4,17 @@ from langchain_core.embeddings import Embeddings
 from langchain_core.tools import BaseTool
 from pydantic import BaseModel, Field, PrivateAttr, model_validator
 
-from databricks_langchain.utils import IndexDetails
 from databricks_langchain.vectorstores import DatabricksVectorSearch
+from databricks_ai_bridge.vector_search_retriever_tool import VectorSearchRetrieverToolMixin, VectorSearchRetrieverToolInput
+from databricks_ai_bridge.utils.vector_search import IndexDetails
 
-
-class VectorSearchRetrieverToolInput(BaseModel):
-    query: str = Field(
-        description="The string used to query the index with and identify the most similar "
-        "vectors and return the associated documents."
-    )
-
-
-class VectorSearchRetrieverTool(BaseTool):
+class VectorSearchRetrieverTool(BaseTool, VectorSearchRetrieverToolMixin):
     """
     A utility class to create a vector search-based retrieval tool for querying indexed embeddings.
-    This class integrates with a Databricks Vector Search and provides a convenient interface
+    This class integrates with Databricks Vector Search and provides a convenient interface
     for building a retriever tool for agents.
     """
 
-    index_name: str = Field(
-        ..., description="The name of the index to use, format: 'catalog.schema.index'."
-    )
-    num_results: int = Field(10, description="The number of results to return.")
-    columns: Optional[List[str]] = Field(
-        None, description="Columns to return when doing the search."
-    )
-    filters: Optional[Dict[str, Any]] = Field(None, description="Filters to apply to the search.")
-    query_type: str = Field(
-        "ANN", description="The type of this query. Supported values are 'ANN' and 'HYBRID'."
-    )
-    tool_name: Optional[str] = Field(None, description="The name of the retrieval tool.")
-    tool_description: Optional[str] = Field(None, description="A description of the tool.")
     text_column: Optional[str] = Field(
         None,
         description="The name of the text column to use for the embeddings. "
@@ -53,7 +33,7 @@ class VectorSearchRetrieverTool(BaseTool):
     _vector_store: DatabricksVectorSearch = PrivateAttr()
 
     @model_validator(mode="after")
-    def validate_tool_inputs(self):
+    def _validate_tool_inputs(self):
         kwargs = {
             "index_name": self.index_name,
             "embedding": self.embedding,
@@ -63,32 +43,8 @@ class VectorSearchRetrieverTool(BaseTool):
         dbvs = DatabricksVectorSearch(**kwargs)
         self._vector_store = dbvs
 
-        def get_tool_description():
-            default_tool_description = (
-                "A vector search-based retrieval tool for querying indexed embeddings."
-            )
-            index_details = IndexDetails(dbvs.index)
-            if index_details.is_delta_sync_index():
-                from databricks.sdk import WorkspaceClient
-
-                source_table = index_details.index_spec.get("source_table", "")
-                w = WorkspaceClient()
-                source_table_comment = w.tables.get(full_name=source_table).comment
-                if source_table_comment:
-                    return (
-                        default_tool_description
-                        + f" The queried index uses the source table {source_table} with the description: "
-                        + source_table_comment
-                    )
-                else:
-                    return (
-                        default_tool_description
-                        + f" The queried index uses the source table {source_table}"
-                    )
-            return default_tool_description
-
         self.name = self.tool_name or self.index_name
-        self.description = self.tool_description or get_tool_description()
+        self.description = self.tool_description or self._get_default_tool_description(IndexDetails(dbvs.index))
 
         return self
