+++ integrations/langchain/tests/unit_tests/test_vectorstores.py
@@ -7,14 +7,16 @@ from databricks.vector_search.client import VectorSearchIndex  # type: ignore
 
 from databricks_langchain.vectorstores import DatabricksVectorSearch
 from tests.utils.vector_search import (
+    EMBEDDING_MODEL,
+    FakeEmbeddings,
+)
+from databricks_ai_bridge.test_utils.vector_search import (  # noqa: F401
     ALL_INDEX_NAMES,
     DELTA_SYNC_INDEX,
     DIRECT_ACCESS_INDEX,
-    EMBEDDING_MODEL,
     ENDPOINT_NAME,
     INDEX_DETAILS,
     INPUT_TEXTS,
-    FakeEmbeddings,
     mock_vs_client,  # noqa: F401
 )
